==> notebooks/00_download_dataset.ipynb <==
# Dataset Download

Downloads the Challenge dataset from google drive if the `data/raw` folder is empty.

## Limitations

- Google Drive folder in the URL should be public.
- Google rate limiters might prevent this code from downloading it. If thats the case, its best to download the data manually and move it to the `data/raw` folder of this project

url = "https://drive.google.com/drive/folders/1Rr5eI8btUAKqDjmDc2vxRyu0C0yRX1Xl"

import os
import gdown

# Target directory
output_dir = os.path.join("..", "data", "raw")
os.makedirs(output_dir, exist_ok=True)

# Check if directory already has files (excluding .gitignore)
existing_files = [f for f in os.listdir(output_dir) if f != ".gitignore"]

if existing_files:
    print(f"⚡ Skipping download, found existing files in {output_dir}")
else:
    print("⬇️  Downloading dataset from Google Drive...")
    gdown.download_folder(url, output=output_dir, quiet=False, use_cookies=False)
    print(f"✅ Download complete. Files saved in: {output_dir}")
